requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
passlib
bcrypt
python-jose
